requirements.txt
pandas
scikit-learn
seaborn
numpy
matplotlib
yfinance
Binance
statsmodels
openpyxl
pysentiment2
selenium
wheel
beautifulsoup4
nltk
sec_edgar_downloader
PyPDF2
faiss-cpu
langchain
openai
alpha_vantage
alpaca_trade_api
datetime
scipy 
unstructured
torch
torchvision
torchaudio
tensorflow
transformers
streamlit
huggingface_hub
python-dotenv
InstructorEmbedding
sentence_transformers
tiktoken
xformers
chromadb
HuggingFaceHub
